Add Event API
This commit adds v2/event_types, v2/events and
v2/events/*/traits. Currently, the event api is
limited to admin users, but is configurable in
the policy.json file. All users that are able to
access the event api can view all events in the
system, it is not restricted per project id.
Operators other than 'eq' have not yet been
implemented.

This change re-names the parameters in the
event_filter to class so they are more
end-user friendly.

implements bp specify-event-api

Change-Id: I0ccd4ef81cac5f8f57a450dc980ea85ff8fe89ec