test code should be excluded from test coverage summary
Since the recent movement of the test code directory, now entire test
code is included in the test coverage summary, it makes a rather
strange loop, test code in the summary also shows being tested or
not tested. This patch will fix the problem so that the test
coverage summary will not show the coverage percentage for test code.

Change-Id: Ibbb4a009f01515befdd7045d6eec1afaf24ca99f