Implement the /v2/samples/<sample-id> API
This is a simple implementation of the new sample endpoint in the API.
It provides raw accesses to the samples collected by Ceilometer, without
imposing any kind of filtering on the meter name.

The patch is about retreiving only one sample with the id.

The fields are also correctly named this time.

Change-Id: Ifdd7f024b066407cd42ec0572385f97acd312125
Blueprint: sample-api