Convert event timestamp to PrecisionTimestamp
Before the timestamps for event and trait are stored in a Float
In mysql, the default precision won't work well, this one are rounded.

This patch, change the Float by the PrecisionTimestamp Column Type, like
the sample timestamp that demand the same precision.

And ensure that get_events always return datetime object.

Change-Id: I1f986784b07afa50bdbb7faa9499e7b614da8ec1