Fix recursive_keypairs output
recursive_keypairs has 'separator' argument. This method is recursive
and a problem is that 'separator' argument is applied only in first
iteration of the method.

Besides, this method is used in flatten_metadata in API implementation.
That's why it is needed not to change output from recursive_keypairs in flatten_metadata.

So the path has a goal to fix incorrect recursive_keypairs impementation but
not to change output of flatten_matadata in API

Fixes bug 1268628

Change-Id: Ic9829342c47325c1df7b81dc4148c08a13e0d067