Make sure use IPv6 sockets for ceilometer in IPv6 environment
In IPv6 management network environment, ceilometer command will
throw 'Address family for hostname not support' when use WSGI
simple server. The root cause is python TCPServer implementation
is hard-coded to use IPv4, even in IPv6 environments. Python
community want to fix it in Python3.5, so we need to make a
workaround for it.

Change-Id: I3260a346a095b0f3e57093c18b1920dffe5ac55d
Closes-Bug: #1269243