Fix the way how metadata is stored in HBase
The goal of this patch is to fix filtering on metadata.
In HBase we should use the same approach as in SQL now:
store raw metadata as well as flattened one.
When user sends get request on meters, samples or resources
it is possible not to fetch all the data but construct corresponding
filter based on flattened metadata values.
Unfortunatelly, it is not possible not to store raw metadata because
flattening on API level is done in a different way. This
improvement is postponed till API v3.

Change-Id: I1206535ab12037817175d5e79716b223594f246a