Change meter_id to meter_name in generated docs
All operations on API level use meter name.
GET or POST /v2/meters/name_of_meter.
In docs we have meter_id, it is confusing, because meter_id is
syntetic UUID that is not used in DB ops.

The change should be done in v2.py because docs are generated
automatically and it's content is based on variables names from
the code.

Change-Id: I29abbefdaaa039f790f78fac102ede933d61d911