nova notifier: disable tests + update sample conf
The nova notifier is again broken due to Nova switch to oslo.messaging.

Anyway, it's likely that this code has been broken for a long time, as
it is not enabled nor tested on devstack anymore. And it's very unlikely
compatible with oslo.messaging.

This patch only disable the run of the tests for now, but we might
consider removing the code before i3 if nobody stands up and try to fix
it in a way or another.

This patch also update the configuration file for the new keystoneclient
release.

Change-Id: I679154baff476957f46e7930db69aeec7e368648