Propogate cacert and insecure flags to glanceclient
Fixes bug 1278973

Previously, the insecure flag was only propagated to the keystoneclient
while the cacert flag was only propagated to the keystoneclient and
the ceilometerclient used by alarming.

Since the glanceclient also supports these options, these flags should
be propogated to it also to avoid SSL cert verification failures when the
central agent calls out to glance over https.

Change-Id: I967ba080a4011a54b5789ca0fc20cb9a37dafb15