Adds time constraints to alarms
This patch allows alarms to have time constraints specified.
If one or more time constraints are specified, the alarm is
evaluated only if the current time is within at least one
time constraint.

A field 'time_constraints' is added to the alarm model
that holds a list of the time constraints. Each time constraint
has the fields
- name		the name of the constraint
- description	description, default is auto-generated from start and duration
- start		starting point(s) of the constraint, in cron format
- duration	duration of the constraint, in seconds
- timezone	optional timezone information

Change-Id: I2d1bcd6728affc31834d7e2f3a0bdd570b2413bb
Blueprint: time-constrained-alarms