Replace hard coded WSGI application creation
This change replaces the hard coded WSGI app creation with a pipeline
of WSGI apps declared in a configuration file.
Paste Deploy was used to create the pipeline since it is used by many other
OpenStack projects and it is an active project
with new contributors and supports Python 3. Dependency on Paste is
localized so switching to another library would not be hard if OpenStack moves
to another package in the future.

The changes are small but the changes for the tests were large since many
acl tests were assuming a hard coded WSGI app creation.

blueprint declarative-filters
Change-Id: I5ce05eab980271873269eca2945dc809f2923045