move databases to test requirements
requirements.txt is only appropriate for libraries that are
always needed, as they will be installed all the time. However
if actually trying to run multinode you won't have databases or
database access on the compute nodes.

So you can't pip install -r requirements.txt on those environments.

For an example, see Nova.

Change-Id: Ic962e518b8ffa0457031d5851f83285abb5096df