Ensure user metadata mapped for instance notifications
Fixes bug 1284073

For pollster-originated samples related to instances, we apply a
mapping to the user metadata so as to ensure that there are no
embedded periods in metadata keys and also to exclude any metadata
not matching the configured reserved namespace.

Now this logic is also applied to user metadata for samples derived
from instance-related notifications in order to avoid failures when
persisting these data in mongodb.

Change-Id: I77eab59023fa23c5879b05c948bd4cb1db7fa177