Enable monkeypatch for select module
When we use ceilometer with kombu (rabbitmq), sometimes when the server
is gone and ceilometer tries to reconnect to it, the connection can hang.

This is because, ceilometer monkeypatch only socket and not select, or
when kombu detect monkeypatching on the socket module, it assume the
select is monkeypatched too and use the real select.select on a eventlet
socket, and hang.

Ceilometer must enable monkeypatch for select too.

Change-Id: Ib41dcde6e5f32755ababf9227827aa8368399c9a
Closes-bug: 1288844