Removes direct access of timeutils.override_time
The attribute "override_time" is a private attribute which
should only be set through "timeutils.set_time_override()".
There are a number of places in Ceilometer which the override_time
attribute is set directly and such behavior should be prohibited.

However the "set_time_override()" function along with "override_time"
in timeutils is planned to be removed, therefore we no longer
need to use it.

Change-Id: I6f3a5e0bb4e4c448f369dd5ead0e1525fd2262e6
Related-Bug: #1266962