Add field translation to complex query from OldSample to Sample
Complex query returns the samples according to the new Sample model, but
it accepts the field names in the query expression only from the OldSample
model during the validation. This patch adds translation between the old
and new field names and translates the field names to the format of the
DB model.

Closes-bug: #1291415

Change-Id: I8ef71b6d43b82cd75f0d7a6ea08cdc8881ca62d3