Use the list when get information from libvirt
If the libvirt adds new elements to the domain info struct or others,
when use the new libvirt, there will be a valueError (too many values
to unpack). When we get information from libvirt, we should use the
original variable format. This change is for compatibility with upcoming
versions of libvirt which change this signature.

Change-Id: I885a48b730acd618e7cefc93cee3d69226575b7c
Closes-Bug: #1291805