VMware vSphere: Improve the accuracy of queried samples
Currently we are querying the latest real time sample from vSphere and
creating a sample out of it, in ceilometer. The sampling interval of
real time samples in vSphere is 20 s. However, the interval at which the
various pollsters are run is by default 600 s (10 mins). This CL
attempts to improve the accuracy of the sample by averaging the samples
over the pollster interval.

Change-Id: Id75fe87840b82d7c7d3e4fae9b2fc982e22ea1d7
Implements: blueprint vmware-vcenter-server
Closes-Bug: 1295500