Verify user/project ID for alarm created by non-admin user
Fixes bug 1297677

Previously, when a non-admin user created an alarm on behalf of
another user, the ownership of the alarm silently reverted to the
requestor's identity.

Now, we check for this case and fail with 401 Not Authorized when
a non-admin user attempts to create an alarm explicitly associated
with another identity.

Change-Id: I8f372b1523012990b8f1d48b03cf9ed9cef65c60