SSL errors thrown with Postgres on multi workers
database connections were made on service initialisation and not
on process start. this led to all processes of a service sharing
the same connection. in this patch, we create unique
dispatch_manager to handle a different engine for each process.

we keep default to a single worker as schema needs to be updated
to avoid possible race/deadlock conditions

Closes-Bug: #1298073
Change-Id: I2ad05e2085c0c0f78653c6354d301d18b8dee121