Align to openstack python package index mirror
Some modules we developed will not be updated to pypi in time, meanwhile,
jenkins will poll packages from openstack mirror, so it is worth to be
set as default index in tox.

User can use -i option in tox cli to change default index server to any
popular or his favorite.

Change-Id: I4449f9b6861a327cf2fd888205843b51a57dcd2f
Closes-Bug: #1289121