Optimize checks to set image properties in metadata
This cleans up the redundant checks we have around
instance.image and ensure we only check if image info
exists once and set the appropriate values and default
to None if missing. Also updated the tests to use the
right default of None instead of empty string and
None assert check for image and new test to validate
image from conductor with no links.

Change-Id: Idb4aa85ae132ec60175ada8a6e304c7eb27f485b