Remove duplicate alarm from alarm_ids
Alarm ids in combination rule is a list, and for now, we allow duplicate
alarms, the duplicate alarms are unnecessary to handle when evaluate a
combination alarm, and it is weird when print the combination rule.

This patch will remove duplicate alarm ids when user try to post or put
combination alarm.

Change-Id: Ic9392355e08834b4d01b0e421de1c77fed8ac5e4
Closes-Bug: #1304886