Update testrepository configuration
This change is a copy of nova testrepository configuration file.

Since we use oslo.test we can now record the stdout and stderr of each
test and print them only is a test fail. This also ensure the printed
log is located into the failing tests.

It also sets a timeout of 160 for each test instead of waiting
indefinitely.

Change-Id: If8e847643d7812a3e595faff171cdc094f5642f4