Downgrade publisher logging to debug level again
https://review.openstack.org/#/c/33448/ changed those strings from
LOG.debug to LOG.audit without a good explanation. I assume that
this was a temporary change to ease debugging. Since LOG.audit
are spamming syslog and this is logged every few seconds, this
is too much churn and it doesn't provide a useful information for
"audit" level. LOG.debug is a saner choice. Also remove
translation.

Change-Id: I772cc578c39f5c84a5fa68f3567aac54ecc03264