Clean tox.ini
The parameters of the documentation related environments
in tox.ini contain redundant and unnecessary parts.

The doc gate job uses the 'venv', which starts a MongoDB
instance. Nor the doc gate job, neither the others like
packaging, which use this env, need a MongoDB instance.
Therefore the 'deps' and the setup-test-env.sh call were
removed from the 'venv' section.

The 'deps' definition was removed from the 'docs' section
also, as it is unnecessary there.

Change-Id: Icfb3e9f88724553cc52143fd24de25f8bcf243a0