Fixes ceilometer-compute service start failure
Currently ceilometer-compute service is failing to start in the
'vmware' region, because of the additional parameter 'port', is
added to the oslo.vmware _get_wsdl_loc method.

This patch fixes it by not calling the private method _get_wsdl_loc,
avoiding the mismatch in the method call. Additional optional
'wsdl_location' is added to over-ride to default location for bug
work-arounds.

Test cases are not added, since the changed method just calls the
oslo.vmware and method is already mocked in the setUp of the
class TestVsphereInspection.

Change-Id: Ib70092c1ccb14421701acdf8ff4a3805becbeef4
Closes-Bug: #1326230