Fix Sphinx directive name in session.py
In openstack/common/db/sqlalchemy/session.py a wrong directive name was
used for formattion code snippets. The directive '.. code::' was changed
to '.. code-block::' in oslo-incubator in this commit:
  I222a11d8a2101600e7cbb96f587515a32f697c2e
As this file now does not contain more formatting errors, it was removed
from the exclude list of Sphinx in conf.py.

Change-Id: I2fc7a8b8216d4bfc5232027f0beb11a036bf257c