Adds caches for image and flavor in compute agent
Currently to obtain image and flavor data,
compute agent always calls nova api, even for the same id of image/flavor.
In order to reduce the number of nova-client call, caches for image and flavor
are introduced.

The caches only live in a local scope of _with_flavor_and_image: created every
single polling interval and thrown away at the end of the interval.
Although the caches only live in a short time, they are effective because same
flavors are used many times in one compute-node in typical use cases.

Change-Id: I63b6a59f06be4911f39073e34cfc3a2a7b832c51
Closes-bug: #1293337