Revert "Align to openstack python package index mirror"
This reverts commit 1c4a7944d003f9d627cece91fa2d2986de2c1d4d.

No other integrated projects are hard-coding the package index.  This
breaks anyone doing CI behind a firewall, and a workaround of -i on the
command line shouldn't be required to use tox with the project.

Changes that require dependencies in the mirror have to go through the
same process as everyone else to get them merged in the
global-requirements file and then served to the mirror so your change
can run with them.

Change-Id: Ifc1e360f8880900a945328f5b62a41e13f9c2df9