Hyper-V: Adds memory metrics implementation
This change targets compute nodes that run on
Windows Hyper-v / Server 2012 or Windows 8 or newer.

It retrieves the memory usage associated with the given
instance and returns a MemoryUsageStats object containing
the value.

This change is particularly useful for compute nodes
that have dynamic memory enabled in nova.conf.

DocImpact

Implements: blueprint hyper-v-memory-metrics

Change-Id: I6496133c3c49ebcc37873004ed58f4d37dbbdd9c