Separate the console scripts
Having all the console scripts in one file means that we always pull
in all the dependencies. This presents a problem in downstream packaging,
for example, a distro alarm subpackage required pymongo dependencies
installed, even if not used.

This patch separates the console scripts into a 'cmd' module, the
way Nova does it. Eventlet monkey patching is done in the module's
__init__.py file.

Co-Authored-By: Mehdi Abaakouk <mehdi.abaakouk@enovance.com>
Closes-bug: #1317210
Change-Id: I83020b13670928fd0d3046d2d2a493b11fe81d61