Replace iterator.next() with next(iterator)
Python 3 introduced a next() function to replace the next()
method on iterator objects. Rather than calling the method on the
iterator, the next() function is called with the iterable object
as it's sole parameter, which calls the underlying __next__() method.

As the next() function was backported to Python 2.6, we can
use it either in the 2.x or 3.x simultaneously - that's why
six.next() usage was also removed from the source code.

Partially-Implements: blueprint ceilometer-py33-support
Change-Id: I47ad69b0e5dc0840f733230d485bb1443cff8e5b