Support for per disk volume measurements
This enables us to track per disk read/write
requests and bytes. The existing samples accumulated
by the incoming data are used to yield needed per
disk samples. Also includes the names for the
assocaited disks as part of the metadata. For per
device pollsters, the resource id is a combination
of instance id and device name. Unit tests included
to cover the above use cases.

Change-Id: I69ab6a99c61e051cfb367c1276b8886f25ae6be7
Implements: blueprint instance-per-disk-measurement