Handle sqlalchemy connection strings with drivers
SqlAlchemy connection strings may specify a driver using the +foo suffix
syntax in the url scheme. Ceilometer should accomodate this type of
schema when looking up sqlalchemy drivers with the stevedore driver
manager.

Fix this by splitting connection string schemes on '+' and retrieving
only the first element in the resulting list. Doing the entrypoint
lookup on this element will preserve the existing behavior in loading
sqlalchemy drivers and should be a noop for other backend types.

Change-Id: Ieec5a4c47e8498a533d4e6954682a9bbe10bd4d9
Fixes-bug: 1359049