Allow pecan debug middleware to be turned off
The Pecan DebugMiddleware is turned on or off based on the
ceilometer.conf-wide debug setting. This is not useful for situations
where running ceilometer-api under mod_wsgi with multiple processes
is desired.

This change adds a pecan_debug setting to the api section of config,
defaulting to the DEFAULT debug value. mod_wsgi install docs updated
accordingly.

Configuration docs have not yet been udpated. They will be pending a
fix to bug 1370030

Closes-Bug: #1370009

Change-Id: Ib4a472e3c8d1f34f64a5f3ab993c1211dee8af9a