Fix neutron client to catch 404 exceptions
When network services such as lbaas, fwaas or vpnaas are disabled
in neutron, the discovery continues to poll the api calls and gets
back a not found exception. The fix here is to catch the exception
so it doesn't go unhandled.

Change-Id: I8f350b9009f0d8c172836b1dd1123e966f887fdb
Closes-Bug: #1374012
(cherry picked from commit b65554eb460a282a2ab0a2dcc0053a8691cb9373)