Implement redesigned separator in names of columns in HBase
This change improves constructing "keys" for stored data in HBase. Now
we are using separators "_!+" to construct the "key", but we don't restrict
any character inside the names of metrics, so it could be a reason of
incorrect parsing of stored data.

In this patch only one separator ":" is used, and method "prepare_key"
helps to avoid incorrect parsing by "quoting" columns and rows names.

Closes-bug: #1328114
Change-Id: I8f816055bb76e67d239de73a57b7cb0baa94dee5