Fix recording failure for system pollster
The parameter "cumulative" was interpreted as a list and could
not be recognized as a valid Type for SQL query hence failing
to record the data for the pollster.

The presence of ',' operator in the pollster class after the term
CUMULATIVE caused it to be treated as a list

This is corrected to interpret the field as a string which will
be successfully recorded to the meter database.

Change-Id: I10a69134a7f0c42a3e6c0d9bb7568e8d8fd2a932
Closes-Bug: 1378742
(cherry picked from commit f8f63d4b15ce68797d6e16943bd85efb19a77752)