Skip to poll and publish when no resources found
Currently, the polling agent will poll and publish even if no resources
found. Although no samples will be published in publishing process, the
polling record will be recorded in log files, this is easy to mislead.
This change skip to publish and record a clear skip-polling hint in
log if no resources found.

Change-Id: Ia0294d3308a1d709006f9cb772f35ea92badcb38