Add alarm_name field to alarm notification
Currently, alarm notification only has the alarm_id which can be used
to identify an alarm. But in some use cases, user may want to use
some customized information to mark the alarm. So this patch added the
alarm_name field into the notification. The name of alarm is tenant
unique and can be defined by user when alarm is created. When the alarm
consumer (e.g. OS::Heat::ScalingPolicy) receives an alarm, it can now
identify the alarm using the customized info stored in alarm_name field.

Change-Id: If9ffcca3f0de8f45483f30bc30ded91dd9d40d77
Closes-Bug: #1399067