Override retry_interval in MongoAutoReconnectTest
Currently, retry_interval is only override when class definition,
that will not work for all unit test cases. This patch fixes this
issue by putting it to setUp method.

Note, retry_interval is 0 can still work for retry algorithm, we
can reduce the time cost to minimum.

Note, DB2 uses same code as MongoDB for connecting, so if MongoDB
works, so can DB2. But they are different in many ways, and this
patch only aims at improving the unit test speed, so DB2 test can
be discussed and added in another patch.

Change-Id: I3e5e15b6d2f503ad7fb81875470259508eff988c
Closes-Bug: #1390680