transform samples only when transformers exist
the pipeline is oddly setup to loop through each sample and
attempt to transform it even when there are no transformers. this
is madness and should be stopped -- this patch does this.

at best, this should improve performance because of needless
looping, at worst, we get rid of misleading 'tranforming sample'
logs.

Change-Id: I80ba2b7775fda03670c7fa556ad5a779c6eefd27