reset listeners on agent refresh
each time an agent joins/leaves group, we kill and refresh the list
of pipeline listeners each agent has. this list previously did not
clear the killed listeners so the list of listeners of an agent
would grow with a bunch of dead listeners. this patch removes
references to the dead listeners.

Change-Id: If4c9e20bca9643d9a30bb1a6c95e252a4ca21bd7
Closes-Bug: #1406405