use default trait type in event list query
Currently, if we request event list api with trait queries, but unset
trait type or set it to '', then we will get 500 and database backend
will fail with KeyError. That is because we use query.type as trait
type, the query.type is checked only when it is set to bool(type) == True,
otherwise, we treat it as string value, but remain type itself unset.

This patch adds a simple check for it, and set it to default type string
when the type is unset or empty.

Change-Id: I503c35cdbc5c8d51dd9b19a314af64ef9fcecf29
Closes-Bug: #1407374